Refactor the gitClone() to use parameterization.
The gitClone() function was running the exec function before on
unsanitized input from the user. This is a serious security
vulnerability due to script injection. By using execFile the user input
is restricted to being a parameter to the git clone function, losing all
control over a shell instance.

Prettier was also run to clean up the spacing within the file.